float type changed to long additional getAudio Name function

+++ app/src/main/java/com/example/tgplayer/model/Audio.kt
@@ -4,5 +4,5 @@ data class Audio(
     val audioSource: String,
     val name: String,
     val thumbnail: String,
-    val length: Float
+    val length: Long
 )
